1 subfolder done, working code

+++ webrtc-build-scripts_8778/android/webrtc/src/webrtc/modules/audio_coding/codecs/pcm16b/include/pcm16b.h
@@ -7,12 +7,15 @@
  *  in the file PATENTS.  All contributing project authors may
  *  be found in the AUTHORS file in the root of the source tree.
  */
+/**********Changed By Rishabh ***********************/ 
 
 #ifndef WEBRTC_MODULES_AUDIO_CODING_CODECS_PCM16B_MAIN_INTERFACE_PCM16B_H_
 #define WEBRTC_MODULES_AUDIO_CODING_CODECS_PCM16B_MAIN_INTERFACE_PCM16B_H_
 /*
  * Define the fixpoint numeric formats
  */
+
+#include <stddef.h>
 
 #include "webrtc/typedefs.h"
 
@@ -36,8 +39,8 @@ extern "C" {
  *                                Always equal to twice the len input parameter.
  */
 
-int16_t WebRtcPcm16b_Encode(const int16_t* speech,
-                            int16_t len,
+size_t WebRtcPcm16b_Encode(const int16_t* speech,
+                           size_t len,
                             uint8_t* encoded);
 
 /****************************************************************************
@@ -55,8 +58,8 @@ int16_t WebRtcPcm16b_Encode(const int16_t* speech,
  * Returned value               : Samples in speech
  */
 
-int16_t WebRtcPcm16b_Decode(const uint8_t* encoded,
-                            int16_t len,
+size_t WebRtcPcm16b_Decode(const uint8_t* encoded,
+                           size_t len,
                             int16_t* speech);
 
 #ifdef __cplusplus
